refactor: rename files from `_modal` to `_dialog`
